+++ AffinityWatcher/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Xml.Linq;
 
 namespace AffinityWatcher
@@ -17,19 +18,22 @@ namespace AffinityWatcher
 
         internal static void Main()
         {
-            var configs = new List<ProcConfig>();
             try
             {
-                ParseConfig(ConfigName, configs);
+                var configs = ParseConfig(ConfigName);
                 PrintConfigStatus(configs);
 
                 var watcher = new ProcessWatcher(configs);
                 watcher.Start();
 
-                while (!Console.KeyAvailable)
+                var task = Task.Run(async () =>
                 {
-                    System.Threading.Thread.Sleep(100);
-                }
+                    while (!Console.KeyAvailable)
+                    {
+                        await Task.Delay(TimeSpan.FromMilliseconds(100)).ConfigureAwait(false);
+                    }
+                });
+                task.Wait();
 
                 watcher.Stop();
             }
@@ -45,19 +49,24 @@ namespace AffinityWatcher
             Console.WriteLine($"Read data for {configs.Count} process(es):");
             foreach (var config in configs)
             {
-                Console.WriteLine("  {0} => {1}", config.ProcessName, config.TargetAffinity);
+                Console.WriteLine($"  {config.ProcessName} => {config.TargetAffinity}");
             }
         }
 
-        internal static void ParseConfig(string filePath, List<ProcConfig> list)
+        /// <summary>
+        /// Parse xml file.
+        /// </summary>
+        /// <param name="filePath"></param>
+        /// <returns></returns>
+        internal static List<ProcConfig> ParseConfig(string filePath)
         {
-            list.Clear();
+            var result = new List<ProcConfig>();
 
             var tree = XElement.Load(filePath);
             var xmlProcListElement = tree.Element("processes");
             if (xmlProcListElement == null)
             {
-                throw new Exception("process list not found");
+                throw new Exception("Process list not found!");
             }
 
             foreach (var xmlProc in xmlProcListElement.Elements().Where(el => el.Name == "process"))
@@ -72,11 +81,13 @@ namespace AffinityWatcher
 
                 if (!long.TryParse(affinityStr, out var affinity))
                 {
-                    throw new Exception($@"Invalid affinity value in process list (name = ""{name}"")");
+                    throw new Exception($@"Invalid affinity value in process list (name: {name})");
                 }
 
-                list.Add(new ProcConfig(name, affinity));
+                result.Add(new ProcConfig(name, affinity));
             }
+
+            return result;
         }
 
         #endregion