+++ CHANGELOG.md
@@ -6,7 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.20] - 2020-10-30
+### Changed
+- Main waiting thread to task
+### Added
+- Search running processes
+
 ## [0.1.10] - 2020-10-29
 ### Added
 - AffinityWatcher project
+  - ProcConfig
+  - ProcessWatcher
+  - Program
 - AffinityWatcher.Tests project
+  - ProcConfigTests
+  - ProcessWatcherTests
+  - ProgramTests
+  - Utils